muita coisa 2.0

+++ lib/models/section.dart
@@ -71,11 +71,13 @@ class Section extends ChangeNotifier {
   Future<void> delete() async {
     await firestoreRef.delete();
     for (final item in items) {
-      try {
-        final ref = storage.refFromURL(item.image as String);
-        await ref.delete();
-        // ignore: empty_catches
-      } catch (e) {}
+      if ((item.image as String).contains('firebase')) {
+        try {
+          final ref = await storage.refFromURL(item.image as String);
+          await ref.delete();
+          // ignore: empty_catches
+        } catch (e) {}
+      }
     }
   }
 
@@ -105,7 +107,8 @@ class Section extends ChangeNotifier {
     }
 
     for (final original in originalItems) {
-      if (!items.contains(original)) {
+      if (!items.contains(original) &&
+          (original.image as String).contains('firebase')) {
         try {
           final ref = storage.refFromURL(original.image as String);
           await ref.delete();
